requirements.txt
pip == 20.0.2
py == 3.7.7
Django == 3.0.6
pillow == 7.1.2
djangorestframework== 3.11.0